float to int
Reverted float values to ints

+++ team-00/src/main/java/ubc/cosc322/COSC322Test.java
@@ -22,9 +22,9 @@ public class COSC322Test extends GamePlayer{
 	private boolean playing = false;
 	private AIPlayer ai = null;
 
-	private final float policyGeneral = 0;
-	private final float policyWin = -1;
-	private final float policyLoss = 1;
+	private final int policyGeneral = 0;
+	private final int policyWin = -1;
+	private final int policyLoss = 1;
 
     private GameClient gameClient = null; 
     private BaseGameGUI gamegui = null;
